requirements.txt
numpy==1.21.5
opencv-python==4.5.4.60
pip==22.0.4
setuptools==49.2.1
python-dotenv==0.19.2


scikit-image==0.19.1
Pillow==8.4.0
rsa==4.8
pyasn1==0.4.8
six==1.16.0
gast==0.4.0
h5py==3.6.0
scipy==1.7.3
keras==2.7.0
tensorflow==2.7.0
wrapt==1.13.3
requests==2.26.0
cachetools==4.2.4
certifi==2021.10.8
urllib3==1.26.7
tifffile==2021.11.2
imageio==2.13.5
packaging==21.3
networkx==2.6.3
Markdown==3.3.6
oauthlib==3.1.1
idna==3.3
Werkzeug==2.0.2
pyparsing==3.0.6
tensorboard==2.7.0
flatbuffers==2.0
astunparse==1.6.3
termcolor==1.1.0
zipp==3.7.0